requirements-whisper.txt
openai-whisper>=20231117
torch>=2.2.0
